Create T2r::IntegrationTest
